refactor(spf): promote SPF behaviors and actor types to top-level
First half of the 4a/4b organizational pivot. The 4b "SPF-framework
integration" code was living under src/media/ (named as if it were
CML-like), conflating the two buckets. This move separates them.

Changes:

- src/media/behaviors/  → src/behaviors/
  (calculate-presentation-duration, load-text-track-cues,
   quality-switching, resolve-presentation, resolve-track,
   sync-preload-attribute — all DOM-free SPF behaviors)

- src/media/actors/     → src/behaviors/actors/
  (text-tracks and text-track-segment-loader — DOM-free actor
   types and the generic injected-parser factory; actors are
   nested under behaviors because they're building blocks that
   serve behaviors, not peers)

- All tests moved alongside their source files.

Scaffolding:
- New subtree tsconfigs at src/behaviors/ and src/behaviors/actors/,
  both DOM-free (lib ES2022 + WebWorker), referencing core / network
  / media / utils.
- Root tsconfig gains both references.
- playback-engines/hls/tsconfig.json adds references to behaviors/
  and behaviors/actors/.
- vitest.config.ts adds a 'behaviors' project (node env, excluding
  the not-yet-created behaviors/dom/ subtrees).

Import paths swept across src/. No behavior change.

Next (second half): eliminate src/dom/ entirely; redistribute
src/dom/media/ → src/media/dom/, src/dom/text/ → src/media/dom/text/,
src/dom/behaviors/ → src/behaviors/dom/, src/dom/actors/ →
src/behaviors/actors/dom/ (and pull source-buffer-actor from
dom/media/ to behaviors/actors/dom/ since it's a SPF actor, not a
CML-like MSE utility).

All green: build (160 files), tests (707), typecheck, workspace
check, lint.
