+++ packages/spf/src/behaviors/calculate-presentation-duration.ts
@@ -1,8 +1,8 @@
-import { effect } from '../../core/signals/effect';
-import type { Signal } from '../../core/signals/primitives';
-import type { AudioTrack, Presentation, VideoTrack } from '../types';
-import { isResolvedTrack } from '../types';
-import { getSelectedTrack } from '../utils/track-selection';
+import { effect } from '../core/signals/effect';
+import type { Signal } from '../core/signals/primitives';
+import type { AudioTrack, Presentation, VideoTrack } from '../media/types';
+import { isResolvedTrack } from '../media/types';
+import { getSelectedTrack } from '../media/utils/track-selection';
 
 export interface PresentationDurationState {
   presentation?: Presentation;
