+++ packages/spf/src/behaviors/tests/calculate-presentation-duration.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
-import { signal } from '../../../core/signals/primitives';
-import type { AudioTrack, Presentation, VideoTrack } from '../../types';
+import { signal } from '../../core/signals/primitives';
+import type { AudioTrack, Presentation, VideoTrack } from '../../media/types';
 import {
   calculatePresentationDuration,
   canCalculateDuration,
