+++ packages/spf/src/behaviors/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../../../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+    "lib": ["ES2022", "WebWorker"],
+    "exactOptionalPropertyTypes": false,
+    "declarationDir": "../../types/behaviors"
+  },
+  "references": [
+    { "path": "../../../utils" },
+    { "path": "../core" },
+    { "path": "../network" },
+    { "path": "../media" },
+    { "path": "./actors" }
+  ],
+  "include": ["./*.ts", "./tests/**/*.ts"]
+}
