+++ tsconfig.json
@@ -22,6 +22,8 @@
     { "path": "packages/spf/src/core" },
     { "path": "packages/spf/src/network" },
     { "path": "packages/spf/src/media" },
+    { "path": "packages/spf/src/behaviors" },
+    { "path": "packages/spf/src/behaviors/actors" },
     { "path": "packages/spf/src/dom" },
     { "path": "packages/spf/src/playback-engines/hls" },
